ec2-spot-asg-handling-interruptions/cdk/requirements.txt
aws-cdk.core

aws-cdk.aws-ec2
aws-cdk.aws-autoscaling
aws-cdk.aws-iam
